add edit article test

+++ app/helpers/categories_helper.rb
@@ -1,5 +1,6 @@
 module CategoriesHelper
   def vote_or_unvote_btn(article)
+    return unless logged_in?
     vote = Vote.find_by(article: article, user: current_user)
     if vote
       link_to('Delete vote!', article_vote_path(id: vote.id, article_id: article.id), method: :delete, class: "btn btn-danger")
